% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashm_rainbow_plot.R
\name{ashm_rainbow_plot}
\alias{ashm_rainbow_plot}
\title{ASHM Rainbow Plot}
\usage{
ashm_rainbow_plot(
  this_maf,
  metadata,
  exclude_classifications,
  drop_unmutated = FALSE,
  classification_column,
  bed,
  region,
  custom_colours,
  hide_ids = TRUE
)
}
\arguments{
\item{this_maf}{Required parameter. A data frame containing mutations (MAF format) within a region of interest.}

\item{metadata}{A data frame with sample_id as a column.}

\item{exclude_classifications}{Optional argument for excluding specific classifications from a metadata file.}

\item{drop_unmutated}{Boolean argument for removing unmutated sample ids in mutated cases.}

\item{classification_column}{The name of the metadata column to use for ordering and colouring samples.}

\item{bed}{Optional data frame specifying the regions to annotate (required columns: start, end, name).}

\item{region}{Genomic region for plotting in bed format.}

\item{custom_colours}{Provide named vector (or named list of vectors) containing custom annotation colours if you do not want to use standartized pallette.}

\item{hide_ids}{Boolean argument, if TRUE, ids will be removed.}
}
\value{
ggplot2 object
}
\description{
Make a rainbow plot of all mutations in a region, ordered and coloured by metadata.
}
\details{
This function creates a rainbow plot for all mutations in a region. Region can either be specified with the `region` parameter,
or the user can provide a maf that has already been subset to the region(s) of interest with `this_maf`.
Regardless, the user has to provide a maf or maf-like object with `this_maf` as an required parameter.
As a third alternative, the regions can also be specified as a bed file with `bed`.
Lastly, this function has a variety of parameters that can be used to further customize the returned plot in many different ways.
Refer to the parameter descriptions, examples as well as the vignettes for more demonstrations how this function can be called.
}
\examples{
#get data
dohh2_maf = GAMBLR.data::sample_data$grch37$maf \%>\% dplyr::filter(Tumor_Sample_Barcode == "DOHH-2")
dohh2_meta = GAMBLR.data::gambl_metadata \%>\% dplyr::filter(sample_id == "DOHH-2")

#build plot
ashm_rainbow_plot(this_maf = dohh2_maf,
                  metadata = dohh2_meta, 
                  region = "chr6:90975034-91066134")

}
