% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_CN_plot.R
\name{circular_CN_plot}
\alias{circular_CN_plot}
\title{Circular overview of copy number data across a cohort}
\usage{
circular_CN_plot(
  pretty_CN_heatmap_output,
  ideogram = TRUE,
  track_height = 0.1,
  min_correlation = 0.35,
  max_neg_correlation = -0.06,
  del_col = "#0000FF80",
  gain_col = "#FF000080",
  calculate_correlations = FALSE,
  return_correlations = FALSE,
  link_transparency = 0.8,
  labelTheseGenes = c("CD58", "TLR2", "MCL1", "CDKN2A", "TMEM30A", "RHOA", "B2M", "PTEN",
    "FAS", "ETV6", "GRB2", "FCGR2B", "CCND3", "CUX1", "MIR17HG", "TFPT", "CD274", "JAK2",
    "CDK14", "BCL6", "EZH2", "HIST1H1E", "REL", "NOL9", "TNFRSF14", "TOX", "TP53", "RB1",
    "TCF4", "HNRNPD", "BCL2", "NFKBIZ", "TNFAIP3", "PRDM1", "CD70", "MYC")
)
}
\arguments{
\item{pretty_CN_heatmap_output}{}

\item{ideogram}{}

\item{labelTheseGenes}{}
}
\description{
Circular overview of copy number data across a cohort
}
\examples{

CN_out = pretty_CN_heatmap(cn_state_matrix=all_states_binned,these_samples_metadata = dlbcl_genome_meta,return_data = T,labelTheseGenes = labelTheseGenes)

circular_CN_plot(CN_out)

}
