% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_report.R
\name{comp_report}
\alias{comp_report}
\title{sample-level SV/SSM/CN reports in PDF}
\usage{
comp_report(
  this_seg,
  this_seg_path,
  this_maf,
  this_maf_path,
  export_individual_plots = FALSE,
  out
)
}
\arguments{
\item{this_seg}{Required parameter. A SEG file loaded into R.}

\item{this_seg_path}{Required, if `this_seg` is not provided. An absolute path to the seg file of interest.}

\item{this_maf}{Required parameter. A MAF file loaded into R.}

\item{this_maf_path}{Required, if `this_maf` is not provided. An absolute path to the maf file of interest.}

\item{export_individual_plots}{Boolean parameter, set to TRUE to export individual plots.}

\item{out}{Path to output folder.}
}
\value{
Nothing.
}
\description{
Construct pdf with sample-level plots, using minimum of arguments
}
\details{
This function runs the complete collection of `fancy_x_plots` on an incoming MAF (`this_maf`) and SEG file (`this_seg`), with default parameters.
The generated plots are put together into a two-page PDF. In addition, it is also possible to export all individual plots.
This can be done by setting `export_individual_plots` to TRUE.
}
\examples{
\dontrun{

#get data for plotting
dohh2_maf = GAMBLR.data::sample_data$grch37$maf \%>\% dplyr::filter(Tumor_Sample_Barcode == "DOHH-2")
dohh2_seg = GAMBLR.data::sample_data$grch37$seg \%>\% dplyr::filter(ID == "DOHH-2")

#create a PDF report for one sample, as well as exporting all individual plots.
comp_report(this_seg = dohh2_seg,
            this_maf = dohh2_maf,
            out = "",
            export_individual_plots = TRUE)
}

}
