% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_number_vaf_plot.R
\name{copy_number_vaf_plot}
\alias{copy_number_vaf_plot}
\title{CN VAF Plot}
\usage{
copy_number_vaf_plot(
  this_maf,
  this_seg,
  just_segments = FALSE,
  coding_only = TRUE,
  one_chrom,
  genes_to_label,
  add_chr_prefix = FALSE,
  plot_title = "My Plot"
)
}
\arguments{
\item{this_maf}{A maf loaded into R. This is a required paraemter.}

\item{this_seg}{A seg loaded into R. This is a required paraemter.}

\item{just_segments}{Specify whether only the segments will be plotted (instead of mutation VAF). Default is FALSE.}

\item{one_chrom}{Subset plot to one chromosome.}

\item{genes_to_label}{Optional. Provide a vector of genes to label (if mutated).}

\item{add_chr_prefix}{If TRUE, "chr" prefix will be added to chr column. Default is FALSE.}

\item{plot_title}{The title for the created plot.}
}
\value{
Nothing
}
\description{
Create a genome-wide copy number plot for one sample and (optionally) display mutation VAF.
}
\details{
This function requires a maf (`this_maf`) and a seg file (`this_seg`) to construct the returned plot.
This plot is visualizing mutation VAFs per default, this can be turned off with setting `just_segments` to TRUE.
This only plots the segments.
}
\examples{
#get data
dohh2_maf = GAMBLR.data::sample_data$grch37$maf \%>\% dplyr::filter(Tumor_Sample_Barcode == "DOHH-2")
dohh2_seg = GAMBLR.data::sample_data$grch37$seg \%>\% dplyr::filter(ID == "DOHH-2")

#build plot
copy_number_vaf_plot(this_maf = dohh2_maf,this_seg = dohh2_seg, plot_title = "DOHH-2")


}
