% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancy_alignment_plot.R
\name{fancy_alignment_plot}
\alias{fancy_alignment_plot}
\title{Plot Alignment Metrics}
\usage{
fancy_alignment_plot(
  collated_results,
  comparison_group,
  add_mean = TRUE,
  add_corrected_coverage = TRUE,
  plot_sub = ""
)
}
\arguments{
\item{collated_results}{Required parameter. A data frame with collated results for the sample IDs of interest. Preferably, the return from `collate_results` in GAMBLR.}

\item{comparison_group}{Optional argument for plotting mean alignment metrics from another group off sample IDs. This should be a data frame with collated results from `collate_results` in GAMBLR, with sample IDs to be compared.}

\item{add_mean}{Set to TRUE to superimpose mean values of plotted variables. Default is TRUE.}

\item{add_corrected_coverage}{Set to TRUE to add corrected coverage for selected samples. Default is TRUE.}

\item{plot_sub}{Optional parameter, add a subtitle to the alignment metric plot.}
}
\value{
A plot as a ggplot object (grob).
}
\description{
Visualize (stacked barplot) genomic read-subsets (metrics) across a selection of samples.
}
\details{
This function is available for plotting relevant alignment metrics (read-subsets) for a selection of samples. Per default, this plot returns the following read-metrics;
total n reads, total n uniquely mapped reads, total n duplicated reads. This plot can also be superimposed with read metrics from additional samples,
allowing for easy comparisons between different sample populations. To run this function, provide a data frame with collated results from `collate_results` in GAMBLR, subset to the sample IDs of interest.
Similarly, if a comparison group is to be superimposed to the returned plot, provide another collated results for these sample IDs (with `collate_results` from GAMBLR).
In addition, this plot can also add additional read-metrics such as mean values for all plotted metrics and corrected coverage. 
To enable these features, simply set `add_mean` and `add_corrected_coverage` to TRUE (default).
}
\examples{
\dontrun{
dlbcl_a_collated = all_collated \%>\% dplyr::filter(BL_subgroup == "DLBCL-A")
dlbcl_b_collated = all_collated \%>\% dplyr::filter(BL_subgroup == "DLBCL-B")

fancy_alignment_plot(collated_results = dlbcl_a_collated, 
                     comparison_group = dlbcl_b_collated, 
                     add_mean = TRUE, 
                     add_corrected_coverage = TRUE, 
                     plot_sub = "Example Plot")
}

}
