% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancy_ideogram.R
\name{fancy_ideogram}
\alias{fancy_ideogram}
\title{genome-wide ideogram annotated with SSM and CN information}
\usage{
fancy_ideogram(
  gene_annotation,
  this_seg,
  this_seg_path = NULL,
  this_maf,
  this_maf_path = NULL,
  plot_title = "Ideogram",
  plot_subtitle = "Genome-wide Ideogram (grch37).",
  intersect_regions,
  include_ssm = TRUE,
  ssm_count = TRUE
)
}
\arguments{
\item{gene_annotation}{Annotate ideogram with a set of genes. These genes can either be specified as a vector of characters or a data frame.}

\item{this_seg}{Parameter with copy number df already loaded into R.}

\item{this_seg_path}{Parameter with path to external cn file.}

\item{this_maf}{Parameter with maf like df already loaded into R.}

\item{this_maf_path}{Parameter with path to external maf like file.}

\item{plot_title}{Title of plot.}

\item{plot_subtitle}{Optional argument for plot subtitle.}

\item{intersect_regions}{Optional parameter for subset variant calls to specific regions. Should be either a vector of characters (chr:start-end) or data frame with regions.}

\item{include_ssm}{Set to TRUE to plot SSMs (dels and ins).}

\item{ssm_count}{Optional parameter to summarize n variants per chromosome, inlcude_ssm must be set to TRUE.}
}
\value{
A plot as a ggplot object (grob).
}
\description{
Generate sample-level ideogram with copy number information, ssm and gene annotations, etc.
}
\details{
This function generates genome-wide ideograms, visualizing SSM data as well as CN segments.
It is also possible to superimpose the plot with gene annotations. Offering a comprehensive overview of all SSM and CN segments of different aneuploidy.
The plotting of SSM can be toggled with setting `include_ssm` to TRUE. If so, it is also possible to count the number of SSMs per chromosome with `ssm_count = TRUE`.
To get data for plotting, there are a few different options available; 
either provide an already loaded seg file with `this_seg` parameter, or an absolute path to a seg file with `this_seg_path`.
Similarly, a maf can be specified (if `include_ssm = TRUE`) with `this_maf` and `this_maf_path`.
For more info on how to run with these parameters, refer to the parameter descriptions.
In order to annotate the ideogram with genes, simply give the `gene_annotations` parameter a set of genes as a vector of characters or a data frame with gene names in the first column.
Another useful parameter for restricting the plotted regions is to call the function with `intersect_regions`.
This parameter takes a vector of characters or a data frame with regions that the plotted calls are restricted to.
}
\examples{
#get data
dohh2_seg = GAMBLR.data::sample_data$grch37$seg \%>\% dplyr::filter(ID == "DOHH-2")

#build plot
fancy_ideogram(this_seg = dohh2_seg,
               include_ssm = FALSE,
               plot_title = "Example Plot")

}
