% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancy_multisamp_ideogram.R
\name{fancy_multisamp_ideogram}
\alias{fancy_multisamp_ideogram}
\title{Genome-wide ideogram (CN segments) for multiple samples.}
\usage{
fancy_multisamp_ideogram(
  this_seg,
  this_seg_path = NULL,
  plot_title = "CN Segments Ideogram",
  plot_sub = "grch37",
  chr_anno_dist = 3,
  chr_select = paste0("chr", c(1:22)),
  include_cn2 = TRUE,
  kompare = FALSE,
  concordance = TRUE
)
}
\arguments{
\item{this_seg}{Parameter with copy number df (seg) already loaded into R.}

\item{this_seg_path}{Parameter with path to external seg file.}

\item{plot_title}{Main title of plot.}

\item{plot_sub}{Subtitle of plot.}

\item{chr_anno_dist}{Optional parameter to adjust chromosome annotations, default value is 3, increase to adjust annotations to left.}

\item{chr_select}{Optional parameter to subset plot to specific chromosomes. Default value is chr1-22.}

\item{include_cn2}{Set to TRUE for plotting CN states == 2. Default is TRUE.}

\item{kompare}{Boolean statement, set to TRUE to call cnvKompare on the selected samples for plotting concordant (or discordant) cn segments across selected chromosomes.}

\item{concordance}{Boolean parameter to be used when kompare = TRUE. Default is TRUE, to plot discordant segments, set the parameter to FALSE.}
}
\value{
A plot as a ggplot object (grob).
}
\description{
Generate ideograms for selected samples, visualizing copy number variation segments.
}
\details{
To create multi-sample ideograms, i.e showing CN segments across multiple samples, this function was created.
This can be used to infer inheritance patterns, hotspots, etc. across multiple samples or the sample ID but for different timepoints.
In addition, this plot can also allow to only plot concordant (or discordant) cn segments between two samples. i.e how two samples differ, or are alike.
The function automatically detects the number of samples provided and sets the plotting parameters accordingly.
The maximum number of samples this plot can deal with is 4. The user must provide either an already laoded seg ffile with `this_seg`, 
or an absolute path to a seg file with `this_seg_path`
In order to only plot segments that are concordant between the selected samples, set `komapre` to TRUE and `concordance` to TRUE.
To instead plot discordant CN segments, set this parameter to FALSE.
}
\examples{
#get data
my_segs = GAMBLR.data::sample_data$grch37$seg \%>\% dplyr::filter(ID \%in\% c("DOHH-2", "02-13135T", "05-17793T"))

#build plot
fancy_multisamp_ideogram(this_seg = my_segs, include_cn2 = TRUE)

}
