% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancy_propcov_plot.R
\name{fancy_propcov_plot}
\alias{fancy_propcov_plot}
\title{Proportional Coverage Plot.}
\usage{
fancy_propcov_plot(collated_results, comparison_samples, plot_subtitle = "")
}
\arguments{
\item{collated_results}{Required parameter. A data frame with collated results for the sample IDs of interest. Preferably, the return from [GAMBLR::collate_results].}

\item{comparison_samples}{Optional parameter, give the function a vector of sample IDs to be compared against the main plotting group.}

\item{plot_subtitle}{Plotting parameter, subtitle of generated plot.}

\item{seq_type}{Selected seq type for incoming QC metrics.}
}
\value{
A plot as a ggplot object (grob).
}
\description{
Visualize proportional coverage (10X and 30X) for selected samples and add comparison group (optional).
}
\details{
Create a highly customizable plot visualizing proportional alignment metrics i.e what proportions of the aligned reads that show 10x and 30x coverage.
This function can also plot the same results for a comparison group of interest, i.e another table with sample IDs. This can be useful for visualizing how certain cohorts/pathologies compares to each other.
To do this one would call the function with `comparison_samples` parameter. For more info on how to use this function, please refer to examples, vignettes (fancy_vignette) and parameter descriptions.
}
\examples{
#get data
dlbcl_a_collated = collate_results() \%>\% dplyr::filter(BL_subgroup == "DLBCL-A")
dlbcl_b_collated = collate_results() \%>\% dplyr::filter(BL_subgroup == "DLBCL-B")

#build plot
fancy_propcov_plot(collated_results = dlbcl_a_collated, comparison_samples = dlbcl_b_collated)

}
