% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancy_proportions_plot.R
\name{fancy_proportions_plot}
\alias{fancy_proportions_plot}
\title{Proportional Metrics Plot.}
\usage{
fancy_proportions_plot(collated_results, plot_subtitle = "")
}
\arguments{
\item{collated_results}{Required parameter. A data frame with collated results for the sample IDs of interest. Preferably, the return from [GAMBLR::collate_results].}

\item{plot_subtitle}{Plotting parameter, subtitle of generated plot.}
}
\value{
A plot as a ggplot object (grob).
}
\description{
This function also provides parameters for easy customization of the plot aesthetics. For example, the subtitle of the plot can easily be controlled with the `plot_subtitle` parameter.
For usage examples and more information, refer to the parameter descriptions and examples in the fancy vignette.
}
\details{
Visualize proportional metrics for selected samples.
}
\examples{
#get data
dlbcl_a_collated = collate_results() \%>\% dplyr::filter(BL_subgroup == "DLBCL-A")

#build plot
fancy_proportions_plot(collated_results = dlbcl_a_collated, plot_subtitle = "Example Plot")  

}
