% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancy_v_chrcount.R
\name{fancy_v_chrcount}
\alias{fancy_v_chrcount}
\title{n variants per chromosome plot.}
\usage{
fancy_v_chrcount(
  this_maf,
  this_maf_path = NULL,
  this_bedpe,
  this_bedpe_path = NULL,
  collated_results,
  ssm = TRUE,
  plot_title = paste0(this_sample_id),
  y_interval = 1,
  hide_legend = FALSE,
  plot_subtitle = "Variant Count Distribution Per Chromosome",
  chr_select = paste0("chr", c(1:22))
)
}
\arguments{
\item{this_maf}{Parameter with maf like df already loaded into R.}

\item{this_maf_path}{Parameter with path to external maf like file.}

\item{this_bedpe}{Parameter with bedpe like df already loaded into R.}

\item{this_bedpe_path}{Parameter with path to external bedpe like file.}

\item{collated_results}{If supplied, the function will superimpose QC metrics to thee returned plot. Preferably the return from [GAMBLR.results::collate_results].}

\item{ssm}{Seet to FALSE to plot SVs instead of SSM. If set to FALSE, this function expects a bedpe with SVs with `this_bedpe` or an absolute path to such a file with `this_bedpe_path`.}

\item{plot_title}{Title of plot (default to sample ID).}

\item{y_interval}{Optional parameter for specifying intervals on y-axis.}

\item{hide_legend}{Set to True to remove legend from plot, default is FALSE.}

\item{plot_subtitle}{Subtitle for created plot.}

\item{chr_select}{vector of chromosomes to be included in plot, defaults to autosomes.}

\item{variant_type_col}{Index of column holding Variant Type (to be used with either maf_data or maf_path).}

\item{chromosome_col}{Index of column holding Chromosome (to be used with either maf_data or maf_path).}
}
\value{
A plot as a ggplot object (grob).
}
\description{
Visualizing variant (SSM or SVs) counts per chromosome.
}
\details{
Takes a maf data frame (or path to a maf), counts the number of variants per chromosome.
Selected chromosomes (`chr_select`) are plotted along the x-axis and the variant counts are represented on the y-axis.
This function can plot both Structural Variants (SV) and Simple Shared Motifs (SSM).
It plots SVs per default and SSM can be added with setting `ssm = TRUE`.
This plot can also be restricted to only show coding mutations. To do so, set `coding_only` to TRUE.
In addition, the returned plot can also be superimposed with a sample-specific mean coverage (from [GAMBLR::collate_results]).
To do so, set `add_qc_metric` to TRUE. A collection of parameters for customizing the returned plot are also available.
e.g `plot_title`, `y_interval`, `hide_legend`, and `plot_subtitle`.
}
