% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancy_v_count.R
\name{fancy_v_count}
\alias{fancy_v_count}
\title{Total n variants count plot.}
\usage{
fancy_v_count(
  this_maf,
  this_maf_path = NULL,
  this_bedpe,
  this_bedpe_path = NULL,
  ssm = TRUE,
  plot_title = "",
  plot_subtitle = "Variant Count For Selected Contigs",
  chr_select = paste0("chr", c(1:22)),
  variant_select = c("DEL", "INS", "DUP"),
  snp_colours = c(SNP = "#2B9971", DNP = "#993F2B", TNP = "#A62656"),
  hide_legend = FALSE,
  log10_y = FALSE
)
}
\arguments{
\item{this_maf}{Parameter with maf like df already loaded into R.}

\item{this_maf_path}{Parameter with path to external maf like file.}

\item{this_bedpe}{Parameter with bedpe like df already loaded into R.}

\item{this_bedpe_path}{Parameter with path to external bedpe like file.}

\item{ssm}{Set to FALSE to get plotting data from get_combined_sv (SVs). Default value is TRUE (plots SSM retrieved from annotate_cn_by_ssm$maf).}

\item{plot_title}{Title of plot (default to sample ID).}

\item{plot_subtitle}{Subtitle for created plot.}

\item{chr_select}{vector of chromosomes to be included in plot, defaults to autosomes.}

\item{variant_select}{Subtypes of SVs to be included in plot, default is DEL, INS and DUP.}

\item{snp_colours}{Optional vector with colours for SNPs (DNP and TNP).}

\item{hide_legend}{Set to True to remove legend from plot, default is FALSE.}

\item{log10_y}{Set to TRUE to force y axis to be in log10.}
}
\value{
A plot as a ggplot object (grob).
}
\description{
Generate a bar plot visualizing total variant (SSM or SVs) count for selected contigs.
}
\details{
This function creates a bar plot showing the total number of variants for a selected sample.
Convenience parameters for restricting the returned plot are available. For example, with `ssm` (Boolean)
you can toggle if the plot will be in respect to SSM (`ssm = TRUE`) or if you wish to count SVs (`ssm = FALSE`).
Lastly, this plotting function also have convenient parameters for customizing the returned plot, 
e.g `plot_title`, `hide_legend`, and`plot_subtitle` and `snp_colours`. 
It is also possible to control what variants are to be counted with `variant_select`. 
Default is deletions, insertions and duplications, c("DEL", "DUP", "INS").
}
\examples{
#get data
dohh2_bedpe = GAMBLR.data::sample_data$grch37$bedpe \%>\% dplyr::filter(tumour_sample_id == "DOHH-2")
dohh2_maf = GAMBLR.data::sample_data$grch37$maf \%>\% dplyr::filter(Tumor_Sample_Barcode == "DOHH-2")

#build plot
fancy_v_count(this_maf = dohh2_maf, this_bedpe = dohh2_bedpe)

}
