% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancy_v_sizedis.R
\name{fancy_v_sizedis}
\alias{fancy_v_sizedis}
\title{Variant size distribution plot}
\usage{
fancy_v_sizedis(
  this_maf,
  this_maf_path = NULL,
  this_bedpe,
  this_bedpe_path = NULL,
  ssm = TRUE,
  plot_title = "",
  plot_subtitle = "Variant Size Distribution",
  scale_value = "width",
  log_10 = TRUE,
  plot_trim = FALSE,
  chr_select = paste0("chr", c(1:22))
)
}
\arguments{
\item{this_maf}{Parameter with maf like df already loaded into R.}

\item{this_maf_path}{Parameter with path to external maf like file.}

\item{this_bedpe}{Parameter with bedpe like df already loaded into R.}

\item{this_bedpe_path}{Parameter with path to external bedpe like file.}

\item{ssm}{Set to FALSE to get plotting data from get_combined_sv (SVs). Default value is TRUE (plots SSM retrieved from annotate_cn_by_ssm$maf).}

\item{plot_title}{Title of plot (default to sample ID).}

\item{plot_subtitle}{Subtitle for created plot.}

\item{scale_value}{Scale type for violin plot, accepted values are "area", "width", and "count", default is "count.}

\item{log_10}{Boolean statement for y-axis, default is TRUE.}

\item{plot_trim}{If TRUE, trim the tails of the violins to the range of the data. If FALSE (default), don't trim the tails.}

\item{chr_select}{vector of chromosomes to be included in plot, defaults to autosomes.}
}
\value{
A plot as a ggplot object (grob).
}
\description{
Generate a violin plot showing variant (SSM or SVs) size distributions for selected contigs.
}
\details{
Function for plotting variant size distributions. This function takes either a dataf frame with a maf like objecct (`this_maf`),
or an absolute pathh to such a file. In addition, if `ssm =. TRUE` the user can also provide a bedpe (`this_bedpe` or `this_bedpe_path`) with SVs to be inlcuded in the returned plot.
For more information on how to run this function, see parameter descriptions and examples.
}
\examples{
#get data
dohh2_bedpe = GAMBLR.data::sample_data$grch37$bedpe \%>\% dplyr::filter(tumour_sample_id == "DOHH-2")
dohh2_maf = GAMBLR.data::sample_data$grch37$maf \%>\% dplyr::filter(Tumor_Sample_Barcode == "DOHH-2")

#build plot
fancy_v_sizedis(this_maf = dohh2_maf, this_bedpe = dohh2_bedpe)

}
