% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAMBLR.viz-package.R
\docType{package}
\name{GAMBLR.viz-package}
\alias{GAMBLR.viz}
\alias{GAMBLR.viz-package}
\title{GAMBLR.viz: Make Plots for Genomic Analysis of Mature B-cell Lymphomas in R}
\description{
This package is part of the Genomic Analysis of Mature B-cell Lymphomas (GAMBL) project developed by the Morin Lab.
}
\author{
\strong{Maintainer}: Vladimir Souza \email{vsouza@bcgsc.ca}

}
\keyword{internal}
