% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_metadata_to_colours.R
\name{map_metadata_to_colours}
\alias{map_metadata_to_colours}
\title{Map Metadata to Colours}
\usage{
map_metadata_to_colours(
  metadataColumns,
  these_samples_metadata,
  column_alias = list(),
  as_vector = TRUE,
  verbose = FALSE,
  annoAlpha = 1
)
}
\arguments{
\item{metadataColumns}{Names of the metadata columns to assign colours for.}

\item{these_samples_metadata}{Metadata for just the samples you need colours for.}

\item{column_alias}{A list of column_names with values indicating what gambl colour scheme they should use (e.g. pos_neg, pathology, lymphgen).}

\item{as_vector}{Boolean statement that is set to TRUE per default.}

\item{verbose}{Set to TRUE to enable verbose mode (debugging messages.}

\item{annoAlpha}{Optional alpha to apply to annotation colours.}
}
\value{
Either a vector or list of colours.
}
\description{
Assign a colour palette to metadata columns automatically and consistently.
}
\details{
Internally called by `plot_sample_circos` in GAMBLR.viz
}
\examples{
#get colours
all_cols = map_metadata_to_colours(metadataColumns = c("lymphgen",
                                                       "pathology",
                                                       "genetic_subgroup"),
                                   these_samples_metadata = GAMBLR.data::gambl_metadata,
                                   column_alias = list("nothing" = "FL"),
                                   as_vector = F)

}
