% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirrored_lollipop_plot.R
\name{mirrored_lollipop_plot}
\alias{mirrored_lollipop_plot}
\title{Mirrored Lollipop Plot.}
\usage{
mirrored_lollipop_plot(
  maf_df1,
  maf_df2,
  gene = NULL,
  plot_title,
  include_silent = FALSE
)
}
\arguments{
\item{maf_df1}{A data frame containing the mutation data from a given cohort or pathology, etc.}

\item{maf_df2}{A data frame containing the mutation data from an other cohort or pathology, etc.}

\item{gene}{The gene symbol to plot.}

\item{plot_title}{Optional, the title of the plot. Default is gene.}

\item{include_silent}{Logical parameter indicating whether to include silent mutations into coding mutations. Default is FALSE.}
}
\value{
A mirrored lollipop plot.
}
\description{
Generates a visually appealing mirrored lollipop plot.
}
\details{
Retrieve two maf data files of a specific sample or a set of samples for comparison. 
A gene of interest can then be visualized with the given maf data files. Silent mutations can 
be visualized setting include_silent to TRUE.
}
\examples{
library(GAMBLR.data)

metadata <- get_gambl_metadata()
metadata1 <- metadata \%>\%
     filter(cohort == "DLBCL_Hilton")
metadata2 <- metadata \%>\%
     filter(cohort == "DLBCL_Thomas")

maf_df1 <- get_ssm_by_samples(
     these_samples_metadata = metadata1
)

maf_df2 <- get_ssm_by_samples(
     these_samples_metadata = metadata2
)

#construct mirrored_lollipop_plot
mirrored_lollipop_result <- mirrored_lollipop_plot(maf_df1, maf_df2, "IGLL5")

}
