% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_hotspot_seq_logo.R
\name{mutation_hotspot_logo}
\alias{mutation_hotspot_logo}
\title{Represent mutation hot spots as a sequence logo}
\usage{
mutation_hotspot_logo(
  maf_df,
  gene_symbol = "MYC",
  hotspot_position = 128750677,
  genome_build,
  pad_length = 20,
  fasta_path,
  include_reference = TRUE,
  include_AA = FALSE,
  group_AA = FALSE,
  annotate_motif = FALSE,
  annotate_only = FALSE,
  return_data = TRUE,
  base_size = 5,
  aa_size = 3,
  text_size = 5,
  include_title = TRUE
)
}
\value{
a named list containing a ggplot object and various processed data
}
\description{
Generates a summary of the frequency of SSMs in a small
region around a mutation hot spot using a "sequence logo"-style stacked plot
}
\examples{
meta = get_gambl_metadata() \%>\% dplyr::filter(seq_type=="genome")
muts = get_ssm_by_regions(regions_list = "2:136875000-136875097",
                                streamlined = F,
                                these_samples_metadata = meta
                                    )


gene_hotspot_logo(maf_df=muts,
                      gene_symbol = "CXCR4",
                      hotspot_position=136875037,
                      include_AA = T,
                      pad_length = 51,
                      annotate_motif=T,
                      base_size = 4)
}
