% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_circos.R
\name{plot_sample_circos}
\alias{plot_sample_circos}
\title{Sample-level Circos Plot}
\usage{
plot_sample_circos(
  this_bedpe,
  this_seg,
  legend_metadata_columns,
  legend_metadata_names = c(),
  this_seq_type = "genome",
  chrom_list,
  label_genes,
  auto_label_sv = FALSE
)
}
\arguments{
\item{this_bedpe}{Required data frame of SVs. If not provided this function will run `get_manta_sv` to retrieve SVs.}

\item{this_seg}{Required data frame of CNVs. If not provided, this function will run `get_sample_cn_segments` to retrieve CNVs.}

\item{legend_metadata_columns}{Column names from metadata}

\item{legend_metadata_names}{List of metadata names to be plotted.}

\item{this_seq_type}{Seq type for returned CN segments. One of "genome" (default) or "capture".}

\item{chrom_list}{List of chromosomes to be plotted. If not stated, chr1-22+X will bes used.}

\item{label_genes}{Gene labels.}

\item{auto_label_sv}{Default is FALSE}

\item{this_maf}{This parameter does not do anything yet. Maybe it was meant to be implemented.}
}
\value{
Nothing
}
\description{
Plot a sample-centric circos overview.
}
\details{
This function lets the user supply already loaded data frames (SV, CNV, SSM) with the `this_bedpe` and`this_seg` parameters.
Convenient Boolean parameters are also available for restricting the plot to specific mutation types (`include_sv` and `include_cnv`).
}
\examples{
#get data
dohh2_seg = GAMBLR.data::sample_data$grch37$seg \%>\% dplyr::filter(ID == "DOHH-2")
dohh2_bedpe = GAMBLR.data::sample_data$grch37$bedpe \%>\% dplyr::filter(tumour_sample_id == "DOHH-2")

#build plot
plot_sample_circos(this_bedpe = dohh2_bedpe,
                   this_seg = dohh2_seg,
                   chrom_list = c("chr2",
                                  "chr3",
                                  "chr8",
                                  "chr14",
                                  "chr18"))

}
