% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_circular_mutation_frequency_heatmap.R
\name{pretty_circular_mutation_frequency_heatmap}
\alias{pretty_circular_mutation_frequency_heatmap}
\title{pretty_circular_mutation_frequency_heatmap}
\usage{
pretty_circular_mutation_frequency_heatmap(
  prettyOncoplot_output,
  cn_status_matrix,
  collated_results,
  these_samples_metadata,
  genes,
  cluster = T,
  keep_these_pathologies,
  min_sample_num = 20,
  col_fun,
  col_theme,
  return_data = FALSE,
  dend_location = "inside",
  clustering_distance_method = "euclidean",
  border = T,
  split_by_type = FALSE,
  rotate_degrees = 0,
  gap.degree = 15,
  show.sector.labels = FALSE,
  label_cex = 0.5,
  rownames_cex = 0.5,
  include_legend = F,
  colour_labels = F,
  label_group = "text",
  label_alpha
)
}
\arguments{
\item{prettyOncoplot_output}{The output of the prettyOncoplot function}

\item{cn_status_matrix}{The output of get_cn_states}

\item{collated_results}{A list of data frames with sample_id as rownames and features as column names}

\item{these_samples_metadata}{A data frame with metadata. Usually the output of [GAMBLR.results::get_gambl_metadata].}

\item{genes}{A vector of genes to label}

\item{cluster}{Whether to perform clustering. Default is TRUE (clustering is performed).}

\item{keep_these_pathologies}{A vector of pathology values to show in the plot. All the remaining rows will be ignored.}

\item{min_sample_num}{Minimum number of samples in a pathology to be considered for the plot. Pathologies with less than this number will be excluded. (20)}

\item{col_fun}{Color function to modify the default color pallette of the heatmap.}

\item{col_theme}{Alternatively, provide the color theme instead of  `col_fun` to change the default colors of the heatmap.}

\item{return_data}{Conditionally return the formatted data used for the plotting. Default is FALSE (only image is plotted and no data is returned).}

\item{dend_location}{Location of the dendrogram. Default is "inside".}

\item{clustering_distance_method}{Clustering method. Default is "euclidean".}

\item{border}{Whether to draw border around heatmap. Default is TRUE (with border).}

\item{split_by_type}{Whether to split the mutations by type. Default is FALSE (no splitting).}

\item{rotate_degrees}{Rotate labels. Default is 0 (no rotation).}

\item{gap.degree}{Gap degree. Default is 15.}

\item{show.sector.labels}{Show labels for each sector of the heatmap. Default is FALSE (no labels).}

\item{label_cex}{Number indicating the amount by which plotting text and symbols should be scaled relative to the default when displaying the labels. Default is 0.5.}

\item{rownames_cex}{Number indicating the amount by which plotting text and symbols should be scaled relative to the default when displaying the rownames. Default is 0.5.}

\item{include_legend}{Whether to include the legend. Default is FALSE (no legend).}

\item{colour_labels}{Optionally color labels. Default is FALSE (no coloring).}

\item{label_group}{How to group the labels. Default is "text".}

\item{label_alpha}{Value from 0 to 1 to control alpha of the label.}
}
\value{
Nothing or a list of data frames (when return_data = TRUE)
}
\description{
pretty_circular_mutation_frequency_heatmap
}
\examples{

all_gambl_meta = get_gambl_metadata() \%>\%
    filter(!seq_type == "mrna") \%>\%
    filter(pathology \%in\% names(get_gambl_colours("pathology")))

all_coding = get_all_coding_ssm(these_samples_metadata = all_gambl_meta)

genes = filter(GAMBLR.data::lymphoma_genes_dlbcl_v_latest,curated==TRUE) \%>\%
    pull(Gene)
genes = unique(c(genes,filter(GAMBLR.data::lymphoma_genes_mcl_v_latest,,curated==TRUE) \%>\% pull(Gene)))
genes = unique(c(genes,filter(GAMBLR.data::lymphoma_genes_bl_v_latest,,curated==TRUE) \%>\% pull(Gene)))
oncoplot_output = prettyOncoplot(all_coding,
                                genes=genes,
                                minMutationPercent = 2,
                                these_samples_metadata = all_gambl_meta,
                                simplify = T,return_inputs = T)

pretty_circular_mutation_frequency_heatmap(prettyOncoplot_output = oncoplot_output,
                                           keep_these_pathologies = c("FL",
                                               "DLBCL",
                                               "PMBCL",
                                               "BL",
                                               "HGBL"))

genes_and_cn_threshs =
    data.frame(gene_id=c("MYC",
    "MIR17HG", "TNFAIP3","TCF4",
    "TNFRSF14","REL","CD274",
    "FAS","HNRNPD","B2M","TP53",
    "CDKN2A","RB1","PTEN","CD70",
    "TOX","RHOA","CD58","MCL1",
    "ETV6","TMEM30A"),
    cn_thresh=c(3,3,1,3,1,3,3,1,1,1,1,1,1,1,1,1,1,1,3,1,1)) \%>\%
    mutate(name=ifelse(cn_thresh>2,paste0(gene_id,"_gain"),paste0(gene_id,"_loss")))

seg_data = get_cn_segments()

cn_status = get_cnv_and_ssm_status(only_cnv="all",
                        these_samples_metadata = all_gambl_meta,
                        genes_and_cn_threshs = genes_and_cn_threshs,
                        adjust_for_ploidy=T)


pretty_circular_mutation_frequency_heatmap(cn_status_matrix = cn_status,
                                           prettyOncoplot_output = oncoplot_output)

sv_collated = GAMBLR.results:::collate_sv_results() \%>\%
    select(sample_id,ends_with("sv"))

NFKBIZ_genome = GAMBLR.results:::collate_nfkbiz_results() \%>\%
    select(sample_id,NFKBIZ_UTR)
NFKBIZ_capture = GAMBLR.results:::collate_nfkbiz_results(seq_type_filter="capture") \%>\%
    select(sample_id,NFKBIZ_UTR)

HNRNPH1_genome = GAMBLR.results:::collate_hnrnph1_mutations() \%>\%
    select(sample_id,HNRNPH1_splice)
HNRNPH1_capture = GAMBLR.results:::collate_hnrnph1_mutations(seq_type_filter="capture") \%>\%
    select(sample_id,HNRNPH1_splice)

NFKBIZ = bind_rows(NFKBIZ_genome,NFKBIZ_capture)
HNRNPH1 = bind_rows(HNRNPH1_genome,HNRNPH1_capture)

pretty_circular_mutation_frequency_heatmap(cn_status_matrix = cn_status,
                                           collated_results = list(sv_collated,
                                               NFKBIZ,
                                               HNRNPH1),
                                           prettyOncoplot_output = oncoplot_output,
                                           these_samples_metadata = all_gambl_meta)


all_states_binned = get_cn_states(n_bins_split=2500,
    missing_data_as_diploid = T,
    seg_data = seg_data)

CN_out = pretty_CN_heatmap(cn_state_matrix=all_states_binned,
    these_samples_metadata = all_gambl_meta,
    hide_annotations = "chromosome",
    scale_by_sample=T,
    return_data = T)

arm_level_events = categorize_CN_events(CN_out)


arm_level_annotated = rownames_to_column(arm_level_events$arm_events_simplified,"sample_id")

pretty_circular_mutation_frequency_heatmap(cn_status_matrix = cn_status,
                                           collated_results = list(sv_collated,
                                               NFKBIZ,
                                               HNRNPH1,
                                               arm_level_annotated),
                                           prettyOncoplot_output = oncoplot_output,
                                           these_samples_metadata = all_gambl_meta)


ashm_freq = get_ashm_count_matrix(
                 regions_bed = dplyr::mutate(GAMBLR.data::grch37_ashm_regions,
                               name = paste(gene, region, sep = "_")),
                 this_seq_type = "genome"
                 )
ashm_freq_collated = mutate(ashm_freq,across(,~ifelse(.x>0,1,0)))

 ashm_freq_collated = ashm_freq_collated[,colSums(ashm_freq_collated) >130]
 ashm_freq_collated = rownames_to_column(ashm_freq_collated,"sample_id")


}
