% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_MAF_summary.R
\name{pretty_MAF_summary}
\alias{pretty_MAF_summary}
\title{Pretty MAF summary}
\usage{
pretty_MAF_summary(
  maf_data,
  these_samples_metadata,
  base_size = 4,
  top_n = 10,
  point_size = 0.5,
  one_per_sample = FALSE,
  gene_sizes,
  returnEverything = FALSE
)
}
\arguments{
\item{maf_data}{A data frame containing the mutation data.}

\item{base_size}{Base size of fonts to use in Morons theme}

\item{top_n}{Number of genes to show in top_n panel (default 10)}

\item{point_size}{Size of points to use in ggbeeswarm plot (default 0.5)}

\item{returnEverything}{Set to TRUE to get the underlying numbers
and all panels as a list}
}
\value{
A plot
}
\description{
Generates a visually appealling alternative to
MAFtools `plotmafSummary``.
}
\details{
This function reproduces the multi-panel MAF summary plot
generated by MAFtools. The colour, plot aesthetics etc utilize GAMBLR colours
and are consistent with other GAMBLR.viz plots such as prettyOncoplot.
}
\examples{
suppressMessages(library(GAMBLR.open))
cat("running examples for pretty_MAF_summary\n")
FL_meta <- GAMBLR.open::get_gambl_metadata() \%>\%
    dplyr::filter(pathology == "FL") \%>\%
    check_and_clean_metadata(.,duplicate_action="keep_first")
FL_coding <- GAMBLR.open::get_coding_ssm(these_samples_metadata = FL_meta)

pretty_MAF_summary(FL_coding)

pretty_MAF_summary(FL_coding, top_n = 22, base_size = 6)

}
