% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettySeuratPlot.R
\name{prettySeuratPlot}
\alias{prettySeuratPlot}
\title{Plot Seurat-style umap plot}
\usage{
prettySeuratPlot(
  umap_data,
  features,
  ncol = 3,
  colours = NULL,
  round_matrix = FALSE
)
}
\arguments{
\item{umap_data}{The output list of the `make_and_annotate_umap` function.
Must contain `df` (a data frame with sample_id, V1, and V2 columns of
the umap coordinates) and `features` (a data frame with feature matrix,
where the rownames are sample identifiers and each column corresponds to
an individual feature).}

\item{features}{A vector of characters specifying the features to plot. The
order in which the features are specified in this argument will be
respected during the facet plotting. Required argument.}

\item{ncol}{Integer specifying how many columns should be in the final facet
plot. Default is 3 (the plots will be arranged in 3 columns).}

\item{colours}{For a custom color palette, provide a named list of colours,
where the item name corresponds to the value from feature matrix and the
item itself is the hex code for the desired colour.}

\item{round_matrix}{Whether to round values in the provided feature matrix.
Default is FALSE.}
}
\value{
plot
}
\description{
This function uses the output of the `make_and_annotate_umap` to
     generate a facet of plots in the single-cell data style of Seurat
     package.
}
\examples{

\dontrun{
gambl_mu_core_all <- make_and_annotate_umap(...)

prettySeuratPlot(
    gambl_mu_core_all,
    features = c(
        "EZH2", "CD79B"
    ),
    ncol = 2
)

prettySeuratPlot(
    gambl_mu_core_all,
    features = c(
        "EZH2", "CD79B"
    ),
    ncol = 2,
    colours = list(
        "0" = "red",
        "2" = "gold"
    )
)

prettySeuratPlot(
    gambl_mu_core_all,
    features = c(
        "EZB_feats", "MCD_feats",
        "ST2_feats", "BN2_feats"
    ),
    ncol = 4,
    round_matrix = TRUE
)

}

}
