% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyStackedOncoplot.R
\name{prettyStackedOncoplot}
\alias{prettyStackedOncoplot}
\title{Pretty stacked oncoplot}
\usage{
prettyStackedOncoplot(
  these_samples_metadata,
  maf_data,
  metadataColumns = c("pathology", "lymphgen"),
  sortByMetadataColumns,
  seg_data,
  sortByPGA = FALSE,
  cn_state_matrix = NULL,
  ashm_matrix,
  regions_bed,
  genes,
  second_oncoplot_genes,
  sortByGenes,
  genes_CN_thresh,
  gene_cn_states,
  secondPlotType = "pretty_CN_heatmap",
  oncoplot_location = "top",
  cluster_samples = FALSE,
  secondPlotArgs,
  oncoplotArgs,
  returnEverything = FALSE,
  plot_width = NULL,
  oncoplotHeight = 6,
  secondPlotHeight = 6,
  verbose = FALSE,
  row_names_side = "right",
  pctFontSize = 0,
  heatmap_legend_side = "bottom"
)
}
\arguments{
\item{these_samples_metadata}{A metadata data frame with sample_id as
a column. The order of sample IDs in the rows of this data frame will
dictate the order of samples in the oncoplot.}

\item{maf_data}{A data frame with maf data that will be used to populate
the oncoplot. Required parameter.}

\item{metadataColumns}{A character vector of column names that specifies
which columns from these_samples_metadata will be displayed below the
oncoplot.}

\item{sortByMetadataColumns}{Optional vector specifying which metadata columns to
sort samples by.}

\item{seg_data}{CNV segment data (only used if `cn_state_matrix` is missing).}

\item{sortByPGA}{If TRUE, sort by PGA.}

\item{cn_state_matrix}{Optional CN state matrix for CN plots/sorting.}

\item{ashm_matrix}{Optional ASHM matrix for `secondPlotType="prettyMutationDensity"`.}

\item{regions_bed}{Optional BED regions for mutation density plot.}

\item{genes}{Genes to include in the oncoplot.}

\item{second_oncoplot_genes}{Optional genes for a second oncoplot when
`secondPlotType="prettyOncoplot"`.}

\item{sortByGenes}{Genes used to order samples by mutation/CN status.}

\item{genes_CN_thresh}{Named numeric vector of CN thresholds for genes.}

\item{gene_cn_states}{Optional data frame or matrix of ad hoc, pre-computed
per-gene CN status: rows named by sample_id, columns named by Hugo gene
symbol, values encoded as -2 (homozygous deletion), -1 (deletion), 0 (no
change), 1 (gain) or 2 (amplification). When supplied, this is used
directly for CN-based sorting and for the CN marks on the oncoplot,
bypassing `genes_CN_thresh`/`get_cnv_and_ssm_status()` and the specific
bin-labeled `cn_state_matrix` format they require. `cn_state_matrix` is
still needed if you also want the accompanying genome-wide CN heatmap
panel; `gene_cn_states` only replaces the per-gene status used for
sorting/annotation.}

\item{secondPlotType}{Defaults to pretty_CN_heatmap, which is currently
the only option tested with this function.}

\item{oncoplot_location}{"top" or "bottom" placement of the oncoplot.}

\item{cluster_samples}{If TRUE, cluster samples by mutation status.}

\item{secondPlotArgs}{List of arguments passed to the secondary plot function.}

\item{oncoplotArgs}{List of arguments passed to the oncoplot.}

\item{returnEverything}{If TRUE, return plots and intermediate matrices.}

\item{plot_width}{Optional plot width (cm) for ComplexHeatmap.}

\item{oncoplotHeight}{Height of the oncoplot (cm).}

\item{secondPlotHeight}{Height of the second plot (cm).}

\item{verbose}{Set to TRUE for ultra chatty mode.}

\item{row_names_side}{Side for row names in ComplexHeatmap.}

\item{pctFontSize}{Font size for percent labels.}

\item{heatmap_legend_side}{Side for heatmap legend.}
}
\value{
A `Heatmap` object when `returnEverything=FALSE`, otherwise a list
  containing plots and intermediate matrices.
}
\description{
Pretty stacked oncoplot
}
\examples{
cat("Running example for function: prettyStackedOncoplot\n")
suppressMessages(
  suppressWarnings({

suppressPackageStartupMessages(library(GAMBLR.open))

# Prepare some metadata
 dlbcl_genome_meta = get_gambl_metadata() \%>\% 
  dplyr::filter(pathology=="DLBCL",
                seq_type=="genome") \%>\%
  check_and_clean_metadata(.,duplicate_action="keep_first")

# Get CN segments for these samples
dlbcl_seg = get_cn_segments(dlbcl_genome_meta)

# Prepare CN matrix
cn_mat = segmented_data_to_cn_matrix(dlbcl_seg,
                                     these = dlbcl_genome_meta,
                                     adjust_for_ploidy = TRUE)

dlbcl_maf = get_coding_ssm(these_samples_metadata = dlbcl_genome_meta)

}))

genes=c("KMT2D","BCL2","CREBBP","EZH2","MYD88","CD79B","TP53",
        "PIM1","CARD11","SGK1","SOCS1",'TET2',"SPEN",
        "ETV6","CD83","B2M","S1PR2","GNA13","BTG1",
        "BTG2","DDX3X","KLHL6","HIST1H1E","TBL1XR1","SMARCA4")

# we will only sort on the mutation status of these
# genes and in this order
sortGenes = c("TP53","KMT2D","BCL2","EZH2","MYD88","CD79B")


 CN_thresh = c("REL"=4,
              "CDKN2A"=1,
              "MIR17HG"=4,
              "TP53"=1,
              "TNFRSF14"=1,
              "TNFAIP3"=1)

# oncoplot on top with column order dicted by the mutation status
# of sortGenes (in the order of the genes appearance in that vector)

suppressMessages(
  suppressWarnings({
prettyStackedOncoplot(these_samples_metadata = dlbcl_genome_meta,  
                        maf_data = dlbcl_maf,
                        metadataColumns = c("pathology","lymphgen"),
                        sortByMetadataColumns = c("pathology","lymphgen"),
                        cn_state_matrix = cn_mat,
                        genes_CN_thresh = CN_thresh,
                        genes = genes,
                        sortByGenes = sortGenes)

}))
# oncoplot on top. Clustering of mutations is used to order the columns.
suppressMessages(
  suppressWarnings({
prettyStackedOncoplot(these_samples_metadata = dlbcl_genome_meta,  
                      maf_data = dlbcl_maf,
                      metadataColumns = c("pathology","lymphgen"),
                      cluster_samples = TRUE,
                      cn_state_matrix = cn_mat,
                      genes_CN_thresh = CN_thresh,
                      genes = genes)

}))
# make a list of arguments for the second (here, the upper) plot
CN_args = list("keep_these_chromosomes"=c("2"),
              "scale_by_sample" = TRUE)
suppressMessages(
  suppressWarnings({

prettyStackedOncoplot(these_samples_metadata = dlbcl_genome_meta,
                     maf_data = dlbcl_maf,
                     sortByGenes = "REL",
                     metadataColumns = c("pathology","lymphgen"),
                     oncoplot_location = "bottom",
                     secondPlotArgs = CN_args,
                     cn_state_matrix = cn_mat,
                     genes_CN_thresh = CN_thresh,
                     genes = genes,
                     oncoplotHeight = 8,
                     secondPlotHeight=3)

}))
# make a list of arguments for the second (here, the upper) plot
CN_args = list(
              "scale_by_sample" = TRUE,
              "hide_these_chromosomes" = "X")

# Specifying sortByGenes automatically ensures those genes
# are sorted by their CN status then mutation status when
# oncoplot_location is "bottom".
# The order in the upper plot restricts the order of the lower plot.

suppressMessages(
  suppressWarnings({
prettyStackedOncoplot(these_samples_metadata = dlbcl_genome_meta,
                     maf_data = dlbcl_maf,
                     sortByGenes = "TP53",
                     metadataColumns = c("pathology","lymphgen"),
                     oncoplot_location = "bottom",
                     secondPlotArgs = CN_args,
                     cn_state_matrix = cn_mat,
                     genes_CN_thresh = CN_thresh,
                     genes = genes,
                     secondPlotHeight=9)
}))

# Provide ad hoc, pre-computed per-gene CN states directly instead of
# deriving them from cn_state_matrix + genes_CN_thresh via
# get_cnv_and_ssm_status(). Encoding: -2 = homozygous deletion,
# -1 = deletion, 0 = no change, 1 = gain, 2 = amplification.
# cn_state_matrix is still supplied here because it drives the
# accompanying genome-wide CN heatmap panel; gene_cn_states only
# replaces genes_CN_thresh for per-gene sorting/annotation.
suppressMessages(
  suppressWarnings({
set.seed(123)
cn_genes = c("TP53","CDKN2A","MYC")
gene_cn_states = matrix(
  sample(c(-2,-1,0,1,2), size = nrow(dlbcl_genome_meta) * length(cn_genes),
         replace = TRUE, prob = c(0.05,0.15,0.6,0.15,0.05)),
  nrow = nrow(dlbcl_genome_meta),
  dimnames = list(dlbcl_genome_meta$sample_id, cn_genes)
)

prettyStackedOncoplot(these_samples_metadata = dlbcl_genome_meta,
                     maf_data = dlbcl_maf,
                     metadataColumns = c("pathology","lymphgen"),
                     sortByGenes = c("TP53","CDKN2A"),
                     cn_state_matrix = cn_mat,
                     gene_cn_states = gene_cn_states,
                     genes = genes)
}))

some_regions = create_bed_data(GAMBLR.data::grch37_ashm_regions,
                              fix_names = "concat",
                              concat_cols = c("gene","region"),sep="-")

suppressMessages(
  suppressWarnings({

simple_ashm_mat <-
 get_ashm_count_matrix(some_regions,
                       these_samples_metadata = dlbcl_genome_meta)



prettyStackedOncoplot(these_samples_metadata = dlbcl_genome_meta,
                     maf_data = dlbcl_maf,
                     regions_bed= some_regions,
                     metadataColumns = c("pathology","lymphgen"),
                     oncoplot_location = "bottom",
                     ashm_matrix = simple_ashm_mat,
                    secondPlotType = "prettyMutationDensity",
                     secondPlotArgs = list("merge_genes"=TRUE,
                                           region_fontsize=3),
                     genes = genes,
                    cluster_samples = TRUE,
                    secondPlotHeight = 9)

}))

}
