% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_to_proteinpaint.R
\name{ssm_to_proteinpaint}
\alias{ssm_to_proteinpaint}
\title{MAF to ProteinPaint table}
\usage{
ssm_to_proteinpaint(maf, print_removed_rows = FALSE)
}
\arguments{
\item{maf}{A data frame in MAF format.}

\item{print_removed_rows}{Boolean value indicating whether only removed rows, *i.e.* 
rows in `maf` that do not contain values for required columns, should be printed. 
Used for checking purpose.}
}
\value{
A data frame.
}
\description{
This function takes a MAF-like data frame and convert it to the right format 
for visualization with ProteinPaint.
}
\details{
For visualization with ProteinPaint, the output data frame must be saved to a 
file and uploaded to https://proteinpaint.stjude.org/. 

For a valid ProteinPaint table for visualization, the MAF table must contain the required 
columns to be converted to the ProteinPaint format. They are: Hugo_Symbol, RefSeq, 
Chromosome, Start_Position, HGVSp_Short, Variant_Classification. Optional columns used 
from the MAF table are: Mutation_Status, t_alt_count, t_depth, n_alt_count, n_depth, 
Tumor_Sample_Barcode, Reference_Allele. In the case of using functions `get_ssm_by_samples`, 
`get_coding_ssm`, or `get_ssm_by_patients` to get your MAF table and required/desired 
columns are missing, consider using parameter `basic_columns = FALSE` to ensure that your 
MAF has all needed columns. Other columns are taken from the metadata table generated 
internally in `ssm_to_proteinpaint`.
}
\examples{
library(GAMBLR.data)

# get maf
my_maf = get_coding_ssm(basic_columns = FALSE)

# convert maf to ProteinPaint format
pp_df = ssm_to_proteinpaint(my_maf)

}
