% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashm_rainbow_plot.R
\name{ashm_rainbow_plot}
\alias{ashm_rainbow_plot}
\title{ASHM Rainbow Plot}
\usage{
ashm_rainbow_plot(
  region,
  maf_data = NULL,
  these_samples_metadata,
  classification_column = "lymphgen",
  exclude_classifications = NULL,
  sortByColumns = "pathology",
  projection = "grch37",
  drop_unmutated = FALSE,
  bed,
  custom_colours,
  hide_ids = TRUE
)
}
\arguments{
\item{region}{Genomic region for plotting as a string in the format "chr:start-end".}

\item{maf_data}{A optional MAF-data data frame containing mutations within the region of 
interest for the samples in `these_samples_metadata`. Ensure it's genome build matches the `projection` parameter.If not supplied, mutations will be retrieved 
from `get_ssm_by_regions` using `these_samples_metadata` and `region` as inputs.}

\item{these_samples_metadata}{A data frame with at least columns `sample_id` and the corresponding 
columns given with `classification_column` and `sortByColumns`.}

\item{classification_column}{The name of the metadata column to use for colouring samples. 
Default: lymphgen.}

\item{exclude_classifications}{Optional argument for excluding specific classifications within
the `classification_column` values.}

\item{sortByColumns}{A vector containing the column names you want to sort samples on. The default
is to sort by pathology then `classification_column`.}

\item{projection}{Plot variants projected to this reference, one of grch37 (default) or hg38. 
Must match the projection of `maf_data` MAF data object if provided.}

\item{drop_unmutated}{Boolean argument for removing unmutated samples. Default: FALSE.}

\item{bed}{Optional data frame specifying the regions to annotate. Minimum required column are
start, end, and name with name being used to label the region.}

\item{custom_colours}{Provide named vector (or named list of vectors) containing custom annotation 
colours if you do not want to use standartized pallette.}

\item{hide_ids}{Boolean argument, if TRUE (default), sample ids will not appear on the plot.}
}
\value{
ggplot2 object
}
\description{
Make a rainbow plot of all mutations in a region, ordered and coloured by metadata.
}
\details{
This function creates a rainbow plot for all mutations in a region specified with 
 the `region` parameter for samples specified in `these_samples_metadata`. If 
 `these_samples_metadata` is not specified, it will use the samples returned by 
 `get_gambl_metadata()` of the appropriate seq types. To highlight and annotate specific 
 regions on the plot, an optional bed file of the regions can be specified with `bed`.
 An optional MAF-data dataframe with mutations in the `region` can be supplied with
 `maf_data`.
}
\examples{
\dontrun{
suppressMessages(library(GAMBLR.open))

# Get lymphgen colours
lymphgen_colours = GAMBLR.helpers::get_gambl_colours("lymphgen")

# Basic usage
this_region = "chr6:90975034-91066134"
this_metadata = get_gambl_metadata()
ashm_rainbow_plot(region = this_region,
                  these_samples_metadata = metadata, 
                  custom_colours = lymphgen_colours)
}

}
