% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_CN_plot.R
\name{categorize_CN_events}
\alias{categorize_CN_events}
\title{Categorize arm-level and chromosomal CNV events}
\usage{
categorize_CN_events(pretty_CN_heatmap_output)
}
\arguments{
\item{pretty_CN_heatmap_output}{}
}
\description{
Categorize arm-level and chromosomal CNV events
}
