% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancy_qc_plot.R
\name{fancy_qc_plot}
\alias{fancy_qc_plot}
\title{Plot Quality Control Metrics.}
\usage{
fancy_qc_plot(
  collated_results,
  comparison_samples,
  fill_by = "pathology",
  labels = c("sample_id", "cohort"),
  interactive = FALSE,
  plot_title = "",
  y_axis_lab = "",
  return_plotdata = FALSE
)
}
\arguments{
\item{collated_results}{Required parameter. A data frame with collated results for the sample IDs of interest. Preferably, the return from [GAMBLR::collate_results].}

\item{comparison_samples}{Optional parameter, give the function a vector of sample IDs to be compared against the main plotting group. Pathology is default.}

\item{fill_by}{Parameter for specifying fill variable for grouped bar plot. Can be any factor from incoming metadata, e.g pathology, cohort, etc.}

\item{labels}{If HTML plot version is rendered, you can specify what labels should be visible when hovering over the dots. Default is sample id and cohort. This parameter expects a vector of charachters.}

\item{interactive}{Boolean parameter for generating interactive plot (HTML). Default is FALSE.}

\item{plot_title}{Plotting parameter, plot title.}

\item{y_axis_lab}{Plotting parameter, label of y-axis.}

\item{return_plotdata}{Optional parameter, if set to TRUE a vector of acceptable data types for plotting will be returned, and nothing else.}
}
\value{
A plot as a ggplot object (grob).
}
\description{
Plot for visualizing QC metrics and allowing for grouping by different metadata columns.
}
\details{
This function is readily available for visualizing a variety of quality control metrics. To get started, the user can easily overview all the available metrics with `return_plotdata = TRUE`.
When this parameter is set to TRUE, a vector of characters will be returned detailing all the, for this plot, available metrics. After deciding what metric to plot, simply give the metric of choice to the `plot_data` parameter.
This function also lets the user provide a data frame with sample IDs to be included in the plot. Optionally, the user can also provide an already filtered metadata table with sample IDs of interest to the `these_samples_metadata`.
If none of the two parameters are supplied, the user can easily restrict the plot to any cohort and/or pathology of their liking. This is done by calling `keep_cohort` and `keep_pathology`.
If these parameters are used, the function will retrieve metadata for all available GAMBL sample IDs and then subset to the specified cohort or pathology.
The layout of the returned plot can also be further customized with `sort_by`. This parameter controls the order in which samples would appear. Similarly, `fill_by` allows the user to control on what factor the plot will be filled by.
In addition, the generated plot can also be returned as an interactive HTML rendering, allowing the user to easily hover over any of the points in the plot and get expanded information on each data point. To toggle this function, set the `interactive` parameter to TRUE.
If an interactive plot is generated, it is also possible to dictate what information should be available in the plotted data points. Default for this parameter is sample ID and cohort.
Sometimes it can also be useful to see how a subset of samples compares to another group; to do this one could call the function with a vector of additional sample IDs given to the `comparison_samples` parameter (see examples for more information).
lastly, the plot can also be configured with custom plot title and axis labels (`plot_title` and `y_axis_lab`). For more information, see examples and parameter descriptions.
}
