% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancy_snv_chrdistplot.R
\name{fancy_snv_chrdistplot}
\alias{fancy_snv_chrdistplot}
\title{n SNVs per chromosome plot}
\usage{
fancy_snv_chrdistplot(
  this_maf,
  this_maf_path = NULL,
  variant_type_col = 10,
  chromosome_col = 5,
  plot_title = paste0(this_sample_id),
  plot_subtitle = "SNV Distribution Per Chromosome",
  chr_select = paste0("chr", c(1:22)),
  include_dnp = FALSE,
  hide_legend = FALSE
)
}
\arguments{
\item{this_maf}{Parameter with maf like df already loaded into R.}

\item{this_maf_path}{Parameter with path to external maf like file.}

\item{variant_type_col}{Index of column holding Variant Type (to be used with either maf_data or maf_path).}

\item{chromosome_col}{Index of column holding Chromosome (to be used with either maf_data or maf_path).}

\item{plot_title}{Title of plot (default to sample ID).}

\item{plot_subtitle}{Subtitle for created plot.}

\item{chr_select}{vector of chromosomes to be included in plot, defaults to autosomes.}

\item{include_dnp}{Optional argument for including DNPs. Default is FALSE.}

\item{hide_legend}{Set to True to remove legend from plot, default is FALSE.}
}
\value{
A plot as a ggplot object (grob).
}
\description{
Visualizing the number of SNVs per chromosome.
}
\details{
This function takes on an already loaded maf-like data frame, or a path to the maf file of interest.
In addition, the user can also give this function a sample ID and the function will run [GAMBLR::assign_cn_to_ssm]
to get data for plotting. If a maf file or data frame is used, the user has the chance to specify what column
that holds the Variant Type information (`variant_type_col`), in addition the user can also specify what column
in the incoming maf that is corresponding to the chromosome annotations. This function also includes useful subsetting
options. For example, `chr_select` allows the user to restrict the plot to specific chromosomes. `include_dnp` is an optional
argument (Boolean) for if variants of this subtype should be included or not. The plot can also be restricted to only
counting coding mutations (`coding_only`). Flat-file and augmented maf options can be toggled with `from_flatfile`
and `use_augmented_maf`. Both are TRUE by default and should rarely be set to FALSE. Lastly, this plotting function
also have convenient parameters for customizing the returned plot, e.g `plot_title`, `y_interval`, `hide_legend`, and`plot_subtitle`.
}
