% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancy_snv_chrdistplot.R
\name{fancy_snv_chrdistplot}
\alias{fancy_snv_chrdistplot}
\title{n SNVs per chromosome plot}
\usage{
fancy_snv_chrdistplot(
  this_maf,
  this_maf_path = NULL,
  variant_type_col = 10,
  chromosome_col = 5,
  plot_title = "",
  plot_subtitle = "SNV Distribution Per Chromosome",
  chr_select = paste0("chr", c(1:22)),
  include_dnp = FALSE,
  hide_legend = FALSE
)
}
\arguments{
\item{this_maf}{Parameter with maf like df already loaded into R.}

\item{this_maf_path}{Parameter with path to external maf like file.}

\item{variant_type_col}{Index of column holding Variant Type (to be used with either maf_data or maf_path).}

\item{chromosome_col}{Index of column holding Chromosome (to be used with either maf_data or maf_path).}

\item{plot_title}{Title of plot (default to sample ID).}

\item{plot_subtitle}{Subtitle for created plot.}

\item{chr_select}{vector of chromosomes to be included in plot, defaults to autosomes.}

\item{include_dnp}{Optional argument for including DNPs. Default is FALSE.}

\item{hide_legend}{Set to True to remove legend from plot, default is FALSE.}
}
\value{
A plot as a ggplot object (grob).
}
\description{
Visualizing the number of SNVs per chromosome.
}
\details{
This function takes on an already loaded maf-like data frame, or a path to the maf file of interest.
This function also includes useful subsetting options. For example, `chr_select` allows the user to restrict the plot to specific chromosomes. `include_dnp` is an optional
argument (Boolean) for if variants of this subtype should be included or not. Lastly, this plotting function
also have convenient parameters for customizing the returned plot, e.g `plot_title`, `hide_legend`, and`plot_subtitle`.
}
\examples{
#get data
dohh2_maf = GAMBLR.data::sample_data$grch37$maf \%>\% dplyr::filter(Tumor_Sample_Barcode == "DOHH-2")

#build plot
fancy_snv_chrdistplot(this_maf = dohh2_maf)

}
