% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancy_v_sizedis.R
\name{fancy_v_sizedis}
\alias{fancy_v_sizedis}
\title{Variant size distribution plot}
\usage{
fancy_v_sizedis(
  this_maf,
  this_maf_path = NULL,
  this_bedpe,
  this_bedpe_path = NULL,
  ssm = TRUE,
  plot_title = paste0(this_sample_id),
  plot_subtitle = "Variant Size Distribution",
  scale_value = "width",
  log_10 = TRUE,
  plot_trim = FALSE,
  chr_select = paste0("chr", c(1:22))
)
}
\arguments{
\item{this_maf}{Parameter with maf like df already loaded into R.}

\item{this_maf_path}{Parameter with path to external maf like file.}

\item{this_bedpe}{Parameter with bedpe like df already loaded into R.}

\item{this_bedpe_path}{Parameter with path to external bedpe like file.}

\item{ssm}{Set to FALSE to get plotting data from get_combined_sv (SVs). Default value is TRUE (plots SSM retrieved from annotate_cn_by_ssm$maf).}

\item{plot_title}{Title of plot (default to sample ID).}

\item{plot_subtitle}{Subtitle for created plot.}

\item{scale_value}{Scale type for violin plot, accepted values are "area", "width", and "count", default is "count.}

\item{log_10}{Boolean statement for y-axis, default is TRUE.}

\item{plot_trim}{If TRUE, trim the tails of the violins to the range of the data. If FALSE (default), don't trim the tails.}

\item{chr_select}{vector of chromosomes to be included in plot, defaults to autosomes.}
}
\value{
A plot as a ggplot object (grob).
}
\description{
Generate a violin plot showing variant (SSM or SVs) size distributions for selected contigs.
}
\details{
Function for plotting variant size distributions. This function takes either a sample ID given to the `this_sample` parameter.
In addition, the function can also accept an already loaded MAF or MAF-like object given to the `maf_data` parameter.
As a third option, the function can also read a maf from disk (provide path to maf with `maf_path`).
A collection of convenient filtering and data subsetting parameters are also available for this function.
For restricting your data (if plotting data retrieved with `this_sample_id`), the user can choose to only plot coding mutations with setting `coding_only` to TRUE.
This plot can also deal with SVs as well as SSM data. To control this, please use the `ssm` parameter. If set to TRUE and if `this_sample` is called,
the function gets data with annotate_cn_by_ssm and if set to FALSE, the function calls `get_combined_sv` to get SV calls for plotting.
If the user calls either `maf_data` or `maf_path`, there are a collection of parameters available for specifying
the relevant columns in the given data frame (`variant_type_col`, `chhromosome_col`, `start_col`, `end_col`). It is also possible to
restrict the returned plot to any given chromosome. This is done with the `chr_select` parameter (default is all autosomes).
In addition, plot aesthetics can also be controlled with `plot_title`, `plot_subtitle`, `scale_value`, `log10`, and `trim`.
For more info on how to run with these parameters, refer to the parameter descriptions.
}
\examples{
#plot SSM size distributions:
fancy_v_sizedis(this_sample_id = "HTMCP-01-06-00422-01A-01D")

}
