% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MutationTimeR.R
\name{plot_MutationTimeR}
\alias{plot_MutationTimeR}
\title{Plot MutationTimeR Results}
\usage{
plot_MutationTimeR(
  this_sample_metadata,
  timed_cna,
  timed_ssm,
  genome_build,
  verbose = FALSE
)
}
\arguments{
\item{this_sample_metadata}{Metadata with one row containing your sample details}

\item{verbose}{Set to TRUE for a chattier experience. Default is FALSE.}

\item{projection}{Genome build projection (e.g., "hg38" or "grch37").}
}
\description{
This function plots the timing information from MutationTimeR for SSMs and CNAs.
}
\examples{
\dontrun{
my_meta = suppressMessages(get_gambl_metadata()) \%>\% 
  dplyr::filter(sample_id=="01-20985T",
               seq_type=="genome")
timed = GAMBLR.results::get_timed_mutations(my_meta,"hg38")
print(timed$SSM)

print(timed$CNA)

all_plots = plot_MutationTimeR(my_meta,timed$CNA,timed$SSM)

all_plots$full

all_plots$minimal
}
}
