% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_circos.R
\name{plot_sample_circos}
\alias{plot_sample_circos}
\title{Sample-level Circos Plot}
\usage{
plot_sample_circos(
  this_sample_id,
  sv_df,
  cnv_df,
  ssm_df,
  include_sv = TRUE,
  include_ssm = FALSE,
  legend_metadata_columns,
  legend_metadata_names = c(),
  include_cnv = TRUE,
  genome_build,
  this_seq_type = "genome",
  chrom_list,
  label_genes,
  auto_label_sv = FALSE,
  link_colour_column = NULL,
  hide_legend = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{this_sample_id}{Sample ID for the sample to plot.}

\item{sv_df}{Optional data frame of SVs. If not provided this function will run `get_manta_sv` to retrieve SVs.}

\item{cnv_df}{Optional data frame of CNVs. If not provided, this function will run `get_sample_cn_segments` to retrieve CNVs.}

\item{ssm_df}{This parameter does not do anything yet. Maybe it was meant to be implemented.}

\item{include_sv}{Default TRUE. (does not do anything yet).}

\item{include_ssm}{Defaul FALSE. (does not do anything yet).}

\item{legend_metadata_columns}{Column names from metadata}

\item{legend_metadata_names}{List of metadata names to be plotted.}

\item{include_cnv}{Default TRUE. (does not do anything yet).}

\item{genome_build}{genome_build to use (this is usually inferred automatically from your data)}

\item{this_seq_type}{Seq type for returned CN segments. One of "genome" (default) or "capture".}

\item{chrom_list}{List of chromosomes to be plotted. If not stated, chr1-22+X will be used.}

\item{label_genes}{Gene labels (df, list or what type?)}

\item{auto_label_sv}{Default is FALSE}

\item{link_colour_column}{Which column from sv_df to use for link colouring (default: NULL)}

\item{hide_legend}{Set to TRUE if you want to suppress the legend. Particularly useful if you are not using GAMBL data/metadata}

\item{verbose}{Set to TRUE for a chattier experience}
}
\value{
Nothing
}
\description{
Plot a sample-centric circos overview.
}
\details{
This function takes a sample ID in the `this_sample_id` parameter.
Optionally, the user can supply already loaded data frames (SV, CNV, SSM) with the `sv_df`, `cnv_df` and `ssm_df` parameters.
Convenient Boolean parameteers are also avaialble for restricting the plot to specific mutation types (`include_sv`, `include_cnv`, and `include_ssm`).
}
\examples{
library(GAMBLR.data)

plot_sample_circos(this_sample_id = "02-13135T",
                   legend_metadata_columns = c("pathology",
                                               "lymphgen",
                                               "COO_consensus",
                                               "DHITsig_consensus"),
                   legend_metadata_names = c("pathology",
                                             "LymphGen",
                                             "COO",
                                             "DHITsig"),
                   chrom_list = c("chr3",
                                  "chr8",
                                  "chr14",
                                  "chr18"))

}
