% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_lollipop_plot.R
\name{pretty_lollipop_plot}
\alias{pretty_lollipop_plot}
\title{Lollipop Plot}
\usage{
pretty_lollipop_plot(
  this_maf,
  gene,
  plot_title,
  plot_theme = "cbioportal",
  out_name = paste0("my_lollipop_plot_", gene)
)
}
\arguments{
\item{this_maf}{Required parameter. A data frame containing the mutation data (from a MAF).}

\item{gene}{Required parameter. The gene symbol to plot.}

\item{plot_title}{Optional (defaults to gene name).}

\item{plot_theme}{Options: cbioportal(default), blue, simple, nature, nature2, ggplot2, and dark.}

\item{out_name}{Optional, set the file name of the plot, if you export it to disk. Default name is my_lollipop_plot_{gene}.}
}
\value{
Nothing.
}
\description{
Generates a visually appealing and interactive lollipop plot.
}
\details{
This function is depending on a modified version of `readMAF` from the `g3viz` package.
Returned plot is interactive, meaning the user can hover over individual points in the plot to reveal more information.
The plot can also be exported in a variety of file formats inside the interactive view of the lollipop plot.
}
\examples{
#get data
dohh2_maf = GAMBLR.data::sample_data$grch37$maf \%>\% dplyr::filter(Tumor_Sample_Barcode == "DOHH-2")

#build plot
pretty_lollipop_plot(this_maf = dohh2_maf,
                     gene = "MYC",
                     plot_title = "Mutation data for MYC",
                     plot_theme = "nature2")

}
