% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_mutual_exclusivity.R
\name{pretty_mutual_exclusivity}
\alias{pretty_mutual_exclusivity}
\title{Pretty mutual exclusivity plot}
\usage{
pretty_mutual_exclusivity(
  maf_data,
  mutmat,
  genes,
  these_samples_metadata,
  q_threshold = 0.1,
  exclude_insignificant_genes = TRUE,
  engine = "ggcorrplot",
  use_MCC = FALSE,
  font_size = 7,
  use_alpha = FALSE,
  clustering_distance = "binary",
  gene_anno_df,
  size_factor = 0.01,
  split,
  return_data = F,
  include_silent = F,
  include_hotspots = F,
  review_hotspots = F,
  bonferroni = FALSE
)
}
\arguments{
\item{maf_data}{}

\item{mutmat}{}

\item{genes}{}

\item{these_samples_metadata}{}

\item{q_threshold}{}

\item{exclude_insignificant_genes}{}

\item{engine}{}

\item{use_MCC}{}

\item{font_size}{}

\item{use_alpha}{}

\item{clustering_distance}{}

\item{gene_anno_df}{}

\item{size_factor}{}

\item{split}{}

\item{return_data}{Default False}

\item{include_silent}{Default False}

\item{include_hotspots}{Default False}

\item{review_hotspots}{Default False}

\item{bonferroni}{Default False}
}
\description{
Pretty mutual exclusivity plot
}
\examples{
dlbcl_meta = get_gambl_metadata() \%>\% dplyr::filter(pathology=="DLBCL",seq_type!="mrna")
all_coding = get_all_coding_ssm(these = dlbcl_meta)

pretty_mutual_exclusivity(maf_data = all_coding, 
                               genes = lg_genes,
                               these=dlbcl_meta,
                               engine = "ComplexHeatmap",
                               font_size = 6,
                               use_alpha = T,
                               clustering_distance="binary",
                               include_hotspots = T)
}
