% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyRainfallPlot.R
\name{prettyRainfallPlot}
\alias{prettyRainfallPlot}
\title{Rainfall Plot}
\usage{
prettyRainfallPlot(
  label_ashm_genes = TRUE,
  projection = "grch37",
  chromosome,
  this_maf,
  this_bedpe,
  this_maf_path,
  zoom_in_region,
  plot_title = "My Plot",
  label_sv = FALSE
)
}
\arguments{
\item{label_ashm_genes}{Boolean argument indicating whether the aSHM regions will be labeled or not.}

\item{projection}{Specify projection (grch37 or hg38) of mutations. Default is grch37.}

\item{chromosome}{Provide one or more chromosomes to plot. The chr prefix can be inconsistent with projection and will be handled.}

\item{this_maf}{Specify custom MAF data frame of mutations.}

\item{this_bedpe}{Specify custom BEDPE data frame of mutations.}

\item{this_maf_path}{Specify path to MAF file if it is not already loaded into data frame.}

\item{zoom_in_region}{Provide a specific region in the format "chromosome:start-end" to zoom in to a specific region.}

\item{plot_title}{The title of the generated plot.}

\item{label_sv}{Boolean argument to specify whether label SVs or not. Only supported if a specific chromosome or zoom in region are specified.}
}
\value{
a ggplot2 plot. Print it using print() or save it using ggsave()
}
\description{
Plot a rainfall plot for one sample. This function takes in the MAF data frame, or path to a custom MAF file.
If non are specified, the SSM will be obtained through GAMBLR directly (with [GAMBLR::get_ssm_by_region]).
}
\details{
Create a sample-level rainfall plot visualizing single nucleotide substitutions mutations for selected chromosomes.
}
\examples{
#get data for one sample
dohh2_maf = GAMBLR.data::sample_data$grch37$maf \%>\% dplyr::filter(Tumor_Sample_Barcode == "DOHH-2")

#build plot for one sample
prettyRainfallPlot(this_maf = dohh2_maf,
                   zoom_in_region = "8:125252796-135253201")

}
